+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletViewModelBase.cs
@@ -3,6 +3,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
+using WalletWasabi.Fluent.ViewModels.NavBar;
 using WalletWasabi.Helpers;
 using WalletWasabi.Wallets;
